Make ui fixes - Add changes to welcome screen (#888) - Fix transparent statusbar
