+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/backup/BackupActivity.kt
@@ -19,8 +19,6 @@ class BackupActivity : BaseActivity(), BottomConfirmAlert.Listener {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        setTransparentStatusBar()
-
         setContentView(R.layout.activity_backup_words)
 
         viewModel = ViewModelProviders.of(this).get(BackupViewModel::class.java)
