+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/guest/GuestActivity.kt
@@ -4,6 +4,7 @@ import android.os.Bundle
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import io.horizontalsystems.bankwallet.BaseActivity
+import io.horizontalsystems.bankwallet.BuildConfig
 import io.horizontalsystems.bankwallet.R
 import io.horizontalsystems.bankwallet.core.setOnSingleClickListener
 import io.horizontalsystems.bankwallet.modules.backup.BackupModule
@@ -12,13 +13,13 @@ import io.horizontalsystems.bankwallet.modules.restore.RestoreModule
 import io.horizontalsystems.bankwallet.viewHelpers.HudHelper
 import kotlinx.android.synthetic.main.activity_add_wallet.*
 
+
 class GuestActivity : BaseActivity() {
 
     private lateinit var viewModel: GuestViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
         setTransparentStatusBar()
 
         setContentView(R.layout.activity_add_wallet)
@@ -40,7 +41,13 @@ class GuestActivity : BaseActivity() {
         })
 
         viewModel.appVersionLiveData.observe(this, Observer { appVersion ->
-            appVersion?.let { textVersion.text = getString(R.string.Guest_Version, it) }
+            appVersion?.let {
+                var version = it
+                if (getString(R.string.is_release) == "false") {
+                    version = "$version (${BuildConfig.VERSION_CODE})"
+                }
+                textVersion.text = getString(R.string.Guest_Version, version)
+            }
         })
 
         buttonCreate.setOnSingleClickListener {
