+++ app/src/main/res/layout/activity_backup_words.xml
@@ -6,6 +6,7 @@
     >
 
     <androidx.core.widget.NestedScrollView
+        android:id="@+id/nestedScrollView"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:layout_marginBottom="10dp"
