+++ app/src/main/res/values/strings.xml
@@ -19,9 +19,9 @@
 
     <string name="Guest_Title">Unstoppable</string>
     <string name="Guest_Subtitle" translatable="false">IN BLOCKCHAIN WE TRUST</string>
-    <string name="Guest_CreateWallet">New Wallet</string>
-    <string name="Guest_RestoreWallet">Restore Wallet</string>
-    <string name="Guest_Version" translatable="false">v.%s</string>
+    <string name="Guest_CreateWallet">Create Wallet</string>
+    <string name="Guest_RestoreWallet">Restore</string>
+    <string name="Guest_Version" translatable="false">version %s</string>
 
     <!--Restore-->
