+++ app/src/main/res/values/styles.xml
@@ -68,6 +68,7 @@
         <item name="android:colorBackground">@color/dark</item>
         <item name="android:navigationBarColor">@color/black</item>
         <item name="android:windowLightStatusBar">false</item>
+        <item name="android:statusBarColor">@android:color/transparent</item>
 
         <item name="colorControlNormal">@color/grey</item>
         <item name="colorControlActivated">@color/yellow_crypto</item>
@@ -139,6 +140,7 @@
         <item name="android:windowBackground">@drawable/light_gradient_background</item>
         <item name="android:navigationBarColor">@color/black</item>
         <item name="android:windowLightStatusBar">true</item>
+        <item name="android:statusBarColor">@android:color/transparent</item>
 
         <item name="colorControlNormal">@color/grey</item>
         <item name="colorControlActivated">@color/yellow_crypto</item>
